Add DeepRetina tests.

+++ src/main/java/com/actelion/research/orbit/imageAnalysis/deeplearning/playground/maskRCNN/AbstractSegmentationSettings.java
@@ -11,7 +11,7 @@ public abstract class AbstractSegmentationSettings {
     private final int trainingImageTileWidth;
     private final int trainingImageTileHeight;
 
-    private final int tileScaleFactor;
+    private final float tileScaleFactor;
 
     /**
      * Setup and store the settings used for abstract segmentation model training and inference.
@@ -20,11 +20,11 @@ public abstract class AbstractSegmentationSettings {
      * @param tileScaleFactor The scale factor that was used to generate the test dataset (e.g. if image tiles were
      *                        size 8192 px, and training images were output at 512 px, then tileScaleFactor=16.
      */
-    public AbstractSegmentationSettings(int imageWidth, int imageHeight, int tileScaleFactor) {
+    public AbstractSegmentationSettings(int imageWidth, int imageHeight, float tileScaleFactor) {
         this.imageWidth = imageWidth;
         this.imageHeight = imageHeight;
-        this.trainingImageTileWidth = imageWidth * tileScaleFactor;
-        this.trainingImageTileHeight = imageHeight * tileScaleFactor;
+        this.trainingImageTileWidth = (int) (imageWidth * tileScaleFactor);
+        this.trainingImageTileHeight = (int) (imageHeight * tileScaleFactor);
         this.tileScaleFactor = tileScaleFactor;
         this.augmentationSettings = new AugmentationSettings();
     }
@@ -60,7 +60,7 @@ public abstract class AbstractSegmentationSettings {
         return trainingImageTileWidth;
     }
 
-    public int getTileScaleFactor() {
+    public float getTileScaleFactor() {
         return tileScaleFactor;
     }
